+++ src/includes/constant.ts
@@ -3,5 +3,4 @@ export class Constant {
 	public static maximumOpacity       = 0.5;
 	public static minimumOpacity       = 1;
 	public static minimumSlideInterval = 0.1;
-	public static slideIntervalUnit    = ["Hour", "Minute", "Second", "MilliSecond"].map((label) => ({ label }));
 }